Finished convert a pdf file in to tts
Using fitz lib to extract texts from a pdf file and
gtts as the tts lib.

Still trying to convert text into bytes so I don't have to
use the memory to save the audio file for every text and then
delete it after reading. But somehow it doesn't work, might try it later.